+++ src/Expr.php
@@ -24,7 +24,6 @@ use Webmozart\Expression\Comparison\KeyNotExists;
 use Webmozart\Expression\Comparison\LessThan;
 use Webmozart\Expression\Comparison\LessThanEqual;
 use Webmozart\Expression\Comparison\Matches;
-use Webmozart\Expression\Comparison\NotEmpty;
 use Webmozart\Expression\Comparison\NotEquals;
 use Webmozart\Expression\Comparison\NotSame;
 use Webmozart\Expression\Comparison\Same;
@@ -248,11 +247,11 @@ class Expr
     /**
      * Check that a value is not empty.
      *
-     * @return NotEmpty The created expression.
+     * @return Not The created expression.
      */
     public static function notEmpty()
     {
-        return new NotEmpty();
+        return new Not(new IsEmpty());
     }
 
     /**