+++ tests/ExprTest.php
@@ -25,7 +25,6 @@ use Webmozart\Expression\Comparison\KeyNotExists;
 use Webmozart\Expression\Comparison\LessThan;
 use Webmozart\Expression\Comparison\LessThanEqual;
 use Webmozart\Expression\Comparison\Matches;
-use Webmozart\Expression\Comparison\NotEmpty;
 use Webmozart\Expression\Comparison\NotEquals;
 use Webmozart\Expression\Comparison\NotSame;
 use Webmozart\Expression\Comparison\Same;
@@ -85,7 +84,7 @@ class ExprTest extends PHPUnit_Framework_TestCase
             array(
                 'notEmpty',
                 array(),
-                new NotEmpty(),
+                new Not(new IsEmpty()),
             ),
             array(
                 'equals',